refactor: StampPaletteEdit -> StampPaletteDetail
