+++ src/views/Settings/StampPaletteTab/StampPaletteDetailTab.vue
